Resolve Excel crash when adding SubTotalFunction
When adding a SubTotalFunction to a Group field, the resulting Excel
file will cause Excel to crash.  It looks like Excel expects the field
to contain the original elements in the fields items along with the
sort type.

Instead of clearing out the fields inner xml and setting it to the sub
function, we should append the sub function to the existing elements.

```xml
<items count="2">
  <item x="0"/>
  <item x="1"/>
</items>
```

becomes

```xml
<items count="3" subtotal="1">
  <item x="0"/>
  <item x="1"/>
  <item t="sum"/>
</items>
```

Signed-off-by: Ken Herner <kherner@navistone.com>